Improve unit tests
Rsyslog has both unit and integration tests. The former test small
parts of the library while the latter test the library from a higher
perspective.

It seems that the unit tests are written in a non-idiomatic way: each
test function would accomodate multiple test scenarios (and
assertions). This is not optimal as it's better to isolate each unit
test to the specific scenario it actually tests.

Let's fix this. In this commit we update the unit tests found under
`structured_data.rs` file to follow the pattern suggested above, where
each unit test actually tests only a single scenario.